Hide option 'Show Thermal Scan' behind a config flag (fixes #14)
The display option 'Show Thermal Scan' is now hidden behind a
configuration flag and can be enabled by setting the environment
variable 'ENABLE_RASTER_OVERLAY' for the shiny process to 'true'. 

For simplification, the option has also been added to the
'docker-compose.yml' file and is set to 'false' by default. To enable
the 'Show Thermal Scan' display option on the Shiny front-end, change
the value of the 'ENABLE_RASTER_OVERLAY' front-end argument to 'true'.
This should only be done on a local machine or a server that can only
accessed from the intranet and not on a server which can be accessed
from the internet.